This is making the model to load correctly.
After discussing in discuss.pytorch.org with @ptrblck this commit
registers the custom variables needed to reconstruct the model. However,
I realized that loading the model and predicting the data points in the
training set is not working. It seems that the state dictionary that is
dumped is not the optimized one?