Models can be saved/loaded.
This commit makes it possible to train a model, and save its state_dict.
Then at loading time one can recreate the neural network and do
predictions. There are three files created and needed:

    1. Model file: .mlchem.
    2. Parameters for neural network class and fingerprints: .params
    3. Scaler object: .scaler.

I think the scaler object file is something I can get rid of later. But
for the moments it is the easiest way of scaling features using sklearn.