MSELoss with 'mean' reduction.
From pytorch:

'mean': the sum of the output will be divided by the number of elements
in the output.

This is exactly what it is mentioned about Loss Functions for
real-valued outputs with AutoEncoders. The loss has to be divided by the
number of elements in the output. With this change, I can reduce RMSE
between outputs and targets as low as 5e-3.